+++ spec-compliance-tests/babel-tests/check-babel-tests.ts
@@ -32,12 +32,9 @@ es2015/yield/basic-without-argument
 es2015/yield/without-argument
 es2018/async-generators/for-await-async-of
 es2020/bigint/decimal-as-property-name
-es2020/export-ns-from/ns-and-named
-es2022/module-string-names/mixed
 estree/class-private-property/flow
-estree/module-string-names/mixed
-experimental/decorators/export-decorated-class
-experimental/decorators/export-default-decorated-class
+experimental/decorators/parenthesized  // Uses obsolete syntax @(a)()
+experimental/decorators/parenthesized-createParenthesizedExpressions  // Uses obsolete syntax @(a)()
 flow/anonymous-function-no-parens-types/good_15
 flow/arrows-in-ternaries/issue-13644
 flow/arrows-in-ternaries/issue-58
@@ -69,7 +66,6 @@ flow/declare-module/4
 flow/declare-module/5
 flow/declare-module/6
 flow/declare-module/9
-flow/module-string-names/mixed
 flow/multiple-declarations/declare-class
 flow/object-types/getter-key-is-keyword
 flow/opaque-type-alias/opaque_subtype_allow_export
@@ -88,7 +84,8 @@ typescript/import/type-asi
 typescript/import/type-equals-require
 `
   .split("\n")
-  .filter((s) => s);
+  .filter((s) => s)
+  .map((s) => s.split(" ")[0]);
 
 interface ResultSummary {
   numPassed: number;